fix: move encode_fewnerd to fewnerd util
